requirements.txt
flask
flask-cors
openai
python-dotenv
numpy
requests
gunicorn